refactor(generator): rename `generator` module to `chunk`
